==> final-nbs/1.creating-multi-classif-labels.ipynb <==
%load_ext autoreload
%autoreload 2

from pathlib import Path
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import os

PROJECT_DIR = os.path.abspath('.')
if PROJECT_DIR.endswith('final-nbs'):
    PROJECT_DIR = os.path.abspath('../')
    os.chdir(PROJECT_DIR)

import cfg
from src.data import get_features_path_from_metadata, join_dataframe_columns
from src import util
from src.data import setup_directories

util.setup_logging()

dirs = setup_directories(cfg.DATA_DIR, create_dirs=True)

def build_multiclassifcation_target(data: pd.DataFrame, target_name: str) -> pd.Series:
    n_act_targets = data[cfg.TARGETS].sum(axis=1)
    only_target = (n_act_targets == 1) & (data[target_name] == 1)
    multiple_targets = (n_act_targets > 1) & (data[target_name] == 1)
    other_targets = (n_act_targets > 0) & (data[target_name] == 0)
    none_targets = (n_act_targets  == 0) & (data[target_name] == 0)
    
    assert (only_target & multiple_targets).sum() == 0
    assert (only_target & other_targets).sum() == 0
    assert (only_target & none_targets).sum() == 0
    
    multiclass_target = np.full(len(data), np.nan)
    multiclass_target[none_targets] = 0
    multiclass_target[other_targets] = 1
    multiclass_target[only_target] = 2
    multiclass_target[multiple_targets] = 3
    
    assert np.isnan(multiclass_target).sum() == 0
    
    return multiclass_target.astype(np.int64)

raw_dir = Path(dirs['raw'])

pd_train_data = pd.read_csv(raw_dir / 'train_labels.csv', index_col='sample_id')
pd_valid_data = pd.read_csv(raw_dir / 'val_labels.csv', index_col='sample_id')

pd_train_data.head()

multiclass_targets = pd.DataFrame(index=pd_train_data.index, columns=pd_train_data.columns)#.add_suffix('_mutliclass')

for target_name in cfg.TARGETS:
    multiclass_targets[target_name] = build_multiclassifcation_target(pd_train_data, target_name)

multiclass_targets = multiclass_targets.add_suffix('_multiclass')

valid_multiclass_targets = pd.DataFrame(index=pd_valid_data.index, columns=pd_valid_data.columns)#.add_suffix('_mutliclass')

for target_name in cfg.TARGETS:
    valid_multiclass_targets[target_name] = build_multiclassifcation_target(pd_valid_data, target_name)

valid_multiclass_targets = valid_multiclass_targets.add_suffix('_multiclass')

multiclass_targets.nunique()

valid_multiclass_targets.nunique()

train_dir = Path(dirs['train'])

multiclass_targets.to_csv(train_dir / 'multiclass.csv', index=True)

valid_dir = Path(dirs['valid'])

valid_multiclass_targets.to_csv(valid_dir / 'multiclass.csv', index=True)
