==> final-nbs/1.sample-id-aggregation.ipynb <==
%load_ext autoreload
%autoreload 2

from pathlib import Path
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import os

PROJECT_DIR = os.path.abspath('.')
if PROJECT_DIR.endswith('final-nbs'):
    PROJECT_DIR = os.path.abspath('../')
    os.chdir(PROJECT_DIR)

import cfg
from src.data import get_features_path_from_metadata, join_dataframe_columns
from src import util
from src.data import setup_directories

util.setup_logging()

dirs = setup_directories(cfg.DATA_DIR, create_dirs=True)

raw_dir = Path(dirs['raw'])

# read metadata
pd_metadata = pd.read_csv(raw_dir / "metadata.csv", index_col="sample_id")
pd_metadata.head()

train_files = get_features_path_from_metadata(pd_metadata.query("split == 'train'"), raw_dir)
valid_files = get_features_path_from_metadata(pd_metadata.query("split == 'val'"), raw_dir)
test_files = get_features_path_from_metadata(pd_metadata.query("split != 'train'"), raw_dir)

len(train_files), len(valid_files), len(test_files)

from src.preprocessing import apply_preprocessing_fn

def compute_weighted_mass(sample_data: pd.DataFrame):
    mz_abun_sum = (sample_data.groupby(['m/z'])['abundance']
                   .sum()
                   .to_frame('abun_sum')
                   .reset_index())
    total_abun = mz_abun_sum['abun_sum'].sum()
    
    return mz_abun_sum[['m/z', 'abun_sum']].prod(axis=1).sum() / total_abun
    
    
def compute_sample_features(sample_data: pd.DataFrame) -> pd.DataFrame:
    output = {}
    # molecular ion
    output['mol_ion_less99'] = int(sample_data['m/z'].max() < 99)
    # abundance sum could tell us about
    output['weighted_mass'] = compute_weighted_mass(sample_data)
    output['max_temp'] = sample_data['temp'].max()
    output['min_temp'] = sample_data['temp'].min()
    output['temp_range'] = output['max_temp'] - output['min_temp']
    return pd.DataFrame([output]).add_prefix('sample_')
    

_train_sample_features = apply_preprocessing_fn(train_files, pd_metadata, processing_fn=compute_sample_features)

train_sample_features = pd.concat(
    _train_sample_features, names=['sample_id', 'dummy_index']
    ).reset_index(level='dummy_index', drop=True)

_valid_sample_features = apply_preprocessing_fn(valid_files, pd_metadata, processing_fn=compute_sample_features)

valid_sample_features = pd.concat(
    _valid_sample_features, names=['sample_id', 'dummy_index']
    ).reset_index(level='dummy_index', drop=True)

_test_sample_features = apply_preprocessing_fn(test_files, pd_metadata, processing_fn=compute_sample_features)

test_sample_features = pd.concat(
    _test_sample_features, names=['sample_id', 'dummy_index']
    ).reset_index(level='dummy_index', drop=True)

train_dir = Path(dirs['train'])
valid_dir = Path(dirs['valid'])
test_dir = Path(dirs['test'])

from sklearn.preprocessing import QuantileTransformer


for feature in train_sample_features.columns.drop('sample_mol_ion_less99'):
    quantile_transformer = QuantileTransformer(n_quantiles=50).fit(train_sample_features[[feature]])
    train_sample_features[[feature]] = quantile_transformer.transform(train_sample_features[[feature]])
    valid_sample_features[[feature]] = quantile_transformer.transform(valid_sample_features[[feature]])
    test_sample_features[[feature]] = quantile_transformer.transform(test_sample_features[[feature]])
    

train_sample_features.head()

valid_sample_features.head()

test_sample_features.head()

train_sample_features.to_csv(train_dir / 'sample_features.csv', index=True)

valid_sample_features.to_csv(valid_dir / 'sample_features.csv', index=True)

test_sample_features.to_csv(test_dir / 'sample_features.csv', index=True)

### analysis 

# pd_train_target = pd.read_csv(train_dir / 'multiclass.csv', index_col='sample_id')
# train_sample_features = train_sample_features.join(pd_train_target)

# pd_valid_target = pd.read_csv(train_dir / 'multiclass.csv', index_col='sample_id')
# valid_sample_features = valid_sample_features.join(pd_valid_target)

# train_sample_features['is_sam_testbed'] = train_sample_features.index.map(pd_metadata['instrument_type']) == 'sam_testbed'

# test_sample_features['is_sam_testbed'] = test_sample_features.index.map(pd_metadata['instrument_type']) == 'sam_testbed'

# train_sample_features.groupby(train_sample_features['sample_max_mz'] < 99)['is_sam_testbed'].sum()

# test_sample_features.groupby(test_sample_features['sample_max_mz'] < 99)['is_sam_testbed'].sum()

# for target_name in pd_train_target.columns:
    
#     display((train_sample_features.groupby(train_sample_features['sample_max_mz'] < 99)
#      [target_name]
#      .value_counts()
#      .to_frame('count')
#      .sort_index() 
#     ))

# train_sample_features.groupby(train_sample_features['sample_max_mz'] < 99)[cfg.TARGETS].mean()

# valid_sample_features.groupby(valid_sample_features['sample_max_mz'] < 99)[cfg.TARGETS].mean()

# train_mass_under99 = train_sample_features[train_sample_features['sample_max_mz'] < 99].copy()

# # none of the max smple < 99 is samtestbed, so we can conclude that is right
# # for the test set, there is 7 observations that are sam testbed which are less than 99

# # it makes sense that if the molecular ion is less than 99 is because the molecule has a small mass
# # so it can not be a molecule with higher mass

# # for the train set and valid set it is the same, chloride can not have a mass of less than 99

# train_sample_features.groupby(train_sample_features['sample_min_mz'] > 1)['is_sam_testbed'].sum()
# # all of them are from sam testbed, so we can conclude anything

# train_sample_features.groupby(train_sample_features['sample_min_mz'] > 1)['is_sam_testbed'].mean()

# test_sample_features.groupby(test_sample_features['sample_min_mz'] > 1)['is_sam_testbed'].sum()
# # all of them are from sam testbed, so we can conclude anything
# # ok, any of the sam tesbed include hydrogen, whyyyyyy?

# (train_sample_features['sample_min_mz'] > 1).agg(('sum', 'mean'))

# (test_sample_features['sample_min_mz'] > 1).agg(('sum', 'mean'))

# test_sample_features['is_sam_testbed'].mean() # ok
