==> final-nbs/3.test-train-lgbm-models.ipynb <==
%load_ext autoreload
%autoreload 2

from pathlib import Path
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import os

PROJECT_DIR = os.path.abspath('.')
if PROJECT_DIR.endswith('final-nbs'):
    PROJECT_DIR = os.path.abspath('../')
    os.chdir(PROJECT_DIR)

import cfg
from src.data import get_features_path_from_metadata, join_dataframe_columns
from src import util
from src.data import setup_directories
util.setup_logging()

dirs = setup_directories(cfg.DATA_DIR, create_dirs=True)

raw_dir = Path(dirs['raw'])
train_dir = Path(dirs['train'])
cv_dir = Path(dirs['cv']['test'])

# read metadata
pd_metadata = pd.read_csv(raw_dir / "metadata.csv", index_col="sample_id")
pd_metadata.head()

# read train labels
pd_train_target = pd.read_csv(raw_dir / 'train_labels.csv', index_col='sample_id')
# read train labels
pd_valid_target = pd.read_csv(raw_dir / 'val_labels.csv', index_col='sample_id')

# read train labels
pd_multclass_target = pd.read_csv(train_dir / 'multiclass.csv', index_col='sample_id')
pd_agg_features = pd.read_csv(train_dir / 'mz_agg_features_drop_correlated.csv', index_col='sample_id')
pd_cluster_features = pd.read_csv(train_dir / 'ae_clusters.csv', index_col='sample_id')
pd_sample_features = pd.read_csv(train_dir / 'sample_features.csv', index_col='sample_id')
pd_features = pd.concat((pd_sample_features, pd_agg_features, pd_cluster_features), axis=1)

valid_dir = Path(dirs['valid'])

# read train labels
pd_valid_multclass_target = pd.read_csv(valid_dir / 'multiclass.csv', index_col='sample_id')
pd_valid_agg_features = pd.read_csv(valid_dir / 'mz_agg_features_drop_correlated.csv', index_col='sample_id')
pd_valid_cluster_features = pd.read_csv(valid_dir / 'ae_clusters.csv', index_col='sample_id')
pd_valid_sample_features = pd.read_csv(valid_dir / 'sample_features.csv', index_col='sample_id')
pd_valid_features = pd.concat((pd_valid_sample_features, pd_valid_agg_features, pd_valid_cluster_features), axis=1)

pd_features = pd_features.append(pd_valid_features)

pd_train_target = pd_train_target.append(pd_valid_target)

pd_multclass_target = pd_multclass_target.append(pd_valid_multclass_target)

feature_names = pd_features.columns.to_list()

from src import util

data = pd.concat((pd_train_target, pd_multclass_target, pd_features), axis=1)

data.head()

from src.model_selection import get_train_test_tuple_from_split
from src.data import get_cv_paths
from sklearn.linear_model import LogisticRegression
from xgboost import XGBRFClassifier
from lightgbm import LGBMClassifier
from sklearn.svm import SVC

model_config = {'model': 'lgbm',
                'parameters':
               {'class_weight': 'balanced',
                'n_estimators': 50, 'colsample_bytree': 0.3}}

from src import train, inference

def train_one_vs_the_rest(data: pd.DataFrame, model_config, feature_names):
    
    models = {}
    
    for target_name in cfg.TARGETS:
        cv_paths = get_cv_paths(cv_dir, target_name)
        multiclass_target_name = f'{target_name}_multiclass'
        models[target_name] = train.train_cv_from_config(data, model_config, feature_names, multiclass_target_name, cv_paths=cv_paths)
    return models

models = train_one_vs_the_rest(data, model_config, feature_names)

def predict_multiclass_fn(model, test_data):
    probs = model.predict_proba(test_data)
    return probs[:, 2:].sum(axis=1)

scores = {}

for target_name in cfg.TARGETS:
    cv_paths = get_cv_paths(cv_dir, target_name)
    model = models[target_name]
    scores[target_name] = inference.cross_validation_inference(data, target_name, model, cv_paths, predict_multiclass_fn)
scores = pd.DataFrame(scores)

cv_scores = scores.mean()

cv_scores

uniform distribution
basalt            0.148178
carbonate         0.105477
chloride          0.165927
iron_oxide        0.194785
oxalate           0.028967
oxychlorine       0.155488
phyllosilicate    0.253738
silicate          0.182060
sulfate           0.189164
sulfide           0.063096
dtype: float64

avg_loss = cv_scores.mean()

avg_loss

test_dir = Path(dirs['test'])

pd_test_agg_features = pd.read_csv(test_dir / 'mz_agg_features.csv', index_col='sample_id')


pd_test_cluster_features = pd.read_csv(test_dir / 'ae_clusters.csv', index_col='sample_id')

test_sample_features = pd.read_csv(test_dir / 'sample_features.csv', index_col='sample_id')
pd_test_features = pd.concat((test_sample_features, pd_test_agg_features, pd_test_cluster_features), axis=1)

pd_test_features.head()

from src.preprocessing import post_processing_prediction

test_yhat = {}

for target_name in cfg.TARGETS:
    target_models = models[target_name]
    target_yhat = np.mean([predict_multiclass_fn(model, pd_test_features) for model in target_models], axis=0)
    target_yhat = post_processing_prediction(pd_test_features, target_name, target_yhat)
    test_yhat[target_name] = target_yhat
test_yhat = pd.DataFrame(test_yhat, index=pd_test_features.index)

train_yhat = {}
for target_name in cfg.TARGETS:
    target_models = models[target_name]
    cv_paths = get_cv_paths(cv_dir, target_name)
    target_yhat = inference.compute_avg_prediction(data, target_models, cv_paths, predict_multiclass_fn)
    train_yhat[target_name] = target_yhat
train_yhat = pd.DataFrame(train_yhat)

sub_dir = Path(dirs['submission'])

sub_dir = sub_dir.joinpath('lgbm', 'test')

sub_dir.mkdir(exist_ok=True, parents=True)

test_yhat.to_csv(sub_dir / 'submission.csv', index=True)

train_yhat.to_csv(sub_dir / 'train.csv', index=True)
