==> final-nbs/4.ensembling.ipynb <==
%load_ext autoreload
%autoreload 2

from pathlib import Path
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import os

PROJECT_DIR = os.path.abspath('.')
if PROJECT_DIR.endswith('final-nbs'):
    PROJECT_DIR = os.path.abspath('../')
    os.chdir(PROJECT_DIR)

import cfg
from src.data import get_features_path_from_metadata, join_dataframe_columns
from src import util
from src.data import setup_directories
util.setup_logging()

dirs = setup_directories(cfg.DATA_DIR, create_dirs=True)

raw_dir = Path(dirs['raw'])
train_dir = Path(dirs['train'])
cv_dir = Path(dirs['cv']['final-validation'])

# read metadata
pd_metadata = pd.read_csv(raw_dir / "metadata.csv", index_col="sample_id")
pd_metadata.head()

sub_dir = Path(dirs['submission'])

train_target = pd.read_csv(raw_dir / 'train_labels.csv', index_col='sample_id')

train_lgbm = pd.read_csv(sub_dir.joinpath('lgbm', 'validation', 'train.csv'), index_col='sample_id')


train_opt_lgbm = pd.read_csv(sub_dir.joinpath('optimized-iteration-lgbm', 'validation', 'train.csv'), index_col='sample_id')

train_svm = pd.read_csv(sub_dir.joinpath('svm', 'validation', 'train.csv'), index_col='sample_id')

train_preds = {'lgbm': train_lgbm, 'svm': train_svm, 'lgbm_opt': train_opt_lgbm}

from src.metrics import compute_scores

valid_target = pd.read_csv(raw_dir / 'val_labels.csv', index_col='sample_id')

valid_lgbm = pd.read_csv(sub_dir.joinpath('lgbm', 'validation', 'submission.csv'), index_col='sample_id')
valid_lgbm = valid_lgbm.loc[valid_target.index]

valid_svm = pd.read_csv(sub_dir.joinpath('svm', 'validation', 'submission.csv'), index_col='sample_id')
valid_svm = valid_svm.loc[valid_target.index]

valid_opt_lgbm = pd.read_csv(sub_dir.joinpath('optimized-iteration-lgbm', 'validation', 'submission.csv'), index_col='sample_id')
valid_opt_lgbm = valid_opt_lgbm.loc[valid_target.index]

valid_lgbm_raw_target = pd.read_csv(sub_dir.joinpath('lgbm-raw-target', 'validation', 'submission.csv'), index_col='sample_id')
valid_lgbm_raw_target = valid_lgbm_raw_target.loc[valid_target.index]

valid_preds = {'lgbm': valid_lgbm, 'svm': valid_svm, 'lgbm_opt': valid_opt_lgbm}

train_scores = {model: compute_scores(train_target.loc[pred.index], pred) for model, pred in train_preds.items()}
train_scores = pd.DataFrame(train_scores).T.sort_values(by=['avg_loss'])

valid_scores = {model: compute_scores(valid_target, pred) for model, pred in valid_preds.items()}
valid_scores = pd.DataFrame(valid_scores).T.sort_values(by=['avg_loss'])

train_scores

valid_scores

joined_scores = train_scores.assign(set_type='train').append(valid_scores.assign(set_type='valid'))

joined_scores = (
    joined_scores
    .reset_index()
    .rename(columns={'index': 'model'})
    .pivot('model', 'set_type')
)

joined_scores.T

opt model makes to much overfit, lets drop it

train_scores

def ensemble_prediction(preds, weights):
    yhat = np.zeros(preds[0].shape)
    
    for i in range(len(preds)):
        yhat += preds[i] * weights[i]
    
    return yhat

ensembles = {
    'lgbm_45_lgbm_opt_45_svm_10': [0.45, 0.45, 0.1],
    'lgbm_40_lgbm_opt_40_svm_20': [0.4, 0.4, 0.2],
    'lgbm_70_svm_30': [0.7, 0., 0.3],
    
    
}

for ensemble_name, weights in ensembles.items():
    train_preds[ensemble_name] = ensemble_prediction([train_lgbm, train_opt_lgbm, train_svm], weights)
    valid_preds[ensemble_name] = ensemble_prediction([valid_lgbm, valid_opt_lgbm, valid_svm], weights)

train_scores = {model: compute_scores(train_target.loc[pred.index], pred) for model, pred in train_preds.items()}
train_scores = pd.DataFrame(train_scores).T.sort_values(by=['avg_loss'])

valid_scores = {model: compute_scores(valid_target, pred) for model, pred in valid_preds.items()}
valid_scores = pd.DataFrame(valid_scores).T.sort_values(by=['avg_loss'])

train_scores

valid_scores

test_lgbm = pd.read_csv(sub_dir.joinpath('lgbm', 'validation', 'submission.csv'), index_col='sample_id')

test_svm = pd.read_csv(sub_dir.joinpath('svm', 'validation', 'submission.csv'), index_col='sample_id')

test_opt_lgbm = pd.read_csv(sub_dir.joinpath('optimized-iteration-lgbm', 'validation', 'submission.csv'), index_col='sample_id')

assert len(np.unique(map(len, [test_lgbm, test_svm, test_opt_lgbm]))) == 1

assert all(test_lgbm.index == test_svm.index)
assert all(test_lgbm.index == test_opt_lgbm.index)

test_preds = {'lgbm': test_lgbm, 'svm': test_svm, 'lgbm_opt': test_opt_lgbm}

for ensemble_name, weights in ensembles.items():
    test_preds[ensemble_name] = ensemble_prediction([test_lgbm, test_opt_lgbm, test_svm], weights)

ensembles

ensemble_dir = sub_dir.joinpath('ensembles-no-validation')
ensemble_dir.mkdir(exist_ok=True, parents=True)

for ensemble_name in ensembles.keys():
    ensemble_pred = test_preds[ensemble_name]
    ensemble_pred.to_csv(ensemble_dir / f'{ensemble_name}.csv', index=True)

!ls {ensemble_dir}
